requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
requests>=2.31
pyarrow>=14.0
pandas>=2.1
google-cloud-storage>=2.14
sqlalchemy>=2.0
pytest>=9.0.3
python-dateutil>=2.8
httpx>=0.28
pytest-timeout>=0.6